Initialize OTEL exporters outside of `instrumentation.*`
to avoid duplicated registration
